remove databinding
