+++ app/src/main/java/com/infomaniak/drive/ui/bottomSheetDialogs/BackgroundSyncPermissionBottomSheetDialog.kt
@@ -18,21 +18,15 @@
 
 package com.infomaniak.drive.ui.bottomSheetDialogs
 
-import android.app.Activity
-import android.content.Context
-import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.infomaniak.drive.BuildConfig
-import com.infomaniak.drive.databinding.FragmentBottomSheetBackgroundSyncBinding
-import com.infomaniak.drive.ui.bottomSheetDialogs.BackgroundSyncPermissionsBottomSheetDialog.BackgroundSyncPermissionsViewModel.Companion.manufacturerWarning
-import com.infomaniak.drive.ui.bottomSheetDialogs.BackgroundSyncPermissionsBottomSheetDialog.BackgroundSyncPermissionsViewModel.Companion.updateShowBatteryOptimizationDialog
+import com.infomaniak.drive.R
+import com.infomaniak.drive.ui.bottomSheetDialogs.BackgroundSyncPermissionsViewModel.Companion.manufacturerWarning
+import com.infomaniak.drive.ui.bottomSheetDialogs.BackgroundSyncPermissionsViewModel.Companion.updateShowBatteryOptimizationDialog
 import com.infomaniak.drive.utils.DrivePermissions
 import com.infomaniak.lib.core.utils.UtilsUi.openUrl
 import kotlinx.android.synthetic.main.fragment_bottom_sheet_background_sync.*
@@ -43,31 +37,38 @@ class BackgroundSyncPermissionsBottomSheetDialog(private val drivePermissions: D
     private val backgroundSyncPermissionsViewModel: BackgroundSyncPermissionsViewModel by viewModels()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        val binding = FragmentBottomSheetBackgroundSyncBinding.inflate(inflater, container, false)
-        binding.lifecycleOwner = viewLifecycleOwner
-        binding.syncPermissionsModel = backgroundSyncPermissionsViewModel
+        return inflater.inflate(R.layout.fragment_bottom_sheet_background_sync, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
         drivePermissions.registerBatteryPermission(this) { hasPermission -> onPermissionGranted(hasPermission) }
         activity?.updateShowBatteryOptimizationDialog(true)
 
         backgroundSyncPermissionsViewModel.apply {
-            shouldBeWhitelisted.value = checkWhitelisted(false)
+            val isWhiteListed = checkWhitelisted(false)
+            allowBackgroundSyncSwitch.isClickable = !isWhiteListed
+            shouldBeWhitelisted.value = isWhiteListed
+
             shouldBeWhitelisted.observe(viewLifecycleOwner) { shouldBeWhitelisted ->
+                allowBackgroundSyncSwitch.apply {
+                    isChecked = shouldBeWhitelisted
+                    isClickable = !shouldBeWhitelisted
+                }
                 if (shouldBeWhitelisted && isWhitelisted.value != true) checkWhitelisted(true)
             }
             hasDoneNecessary.observe(viewLifecycleOwner) { hasDoneNecessary ->
                 val mustShowBatteryDialog = !(hasDoneNecessary && isWhitelisted.value == true)
                 activity?.updateShowBatteryOptimizationDialog(mustShowBatteryDialog)
             }
+
+            hasDoneNecessaryCheckbox.setOnCheckedChangeListener { _, isChecked -> hasDoneNecessary.value = isChecked }
+            allowBackgroundSyncSwitch.setOnCheckedChangeListener { _, isChecked -> shouldBeWhitelisted.value = isChecked }
         }
 
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
         actionButton.setOnClickListener { dismiss() }
-        openGuideButton.setOnClickListener { context?.openUrl(BuildConfig.SUPPORT_DONT_KILL_MY_APP_URL) }
+        openGuideButton.setOnClickListener { context?.openUrl(SUPPORT_FAQ_BACKGROUND_FUNCTIONING_URL) }
     }
 
     private fun onPermissionGranted(hasPermission: Boolean) {
@@ -80,73 +81,18 @@ class BackgroundSyncPermissionsBottomSheetDialog(private val drivePermissions: D
 
     private fun checkWhitelisted(requestPermission: Boolean): Boolean {
         return drivePermissions.checkBatteryLifePermission(requestPermission).also { hasPermission ->
-            backgroundSyncPermissionsViewModel.showManufacturerWarning(hasPermission)
+            showManufacturerWarning(hasPermission)
         }
     }
 
-    class BackgroundSyncPermissionsViewModel : ViewModel() {
-
-        companion object {
-
-            private const val SHARED_PREFS = "HINT_BATTERY_OPTIMIZATIONS"
-            private const val FLAG_SHOW_BATTERY_OPTIMIZATION_DIALOG = "showBatteryOptimizationDialog"
-
-            /**
-             * List of manufacturers which are known to restrict background processes or otherwise
-             * block synchronization.
-             *
-             * See https://github.com/jaredrummler/AndroidDeviceNames/blob/master/json/ for manufacturer values.
-             */
-            private val evilManufacturers = arrayOf("huawei", "oneplus", "samsung", "xiaomi")
-
-            /**
-             * Whether the device has been produced by an evil manufacturer.
-             * Always true for debug builds (to test the UI).
-             *
-             * @see evilManufacturers
-             */
-            val manufacturerWarning = evilManufacturers.contains(Build.MANUFACTURER.lowercase()) || BuildConfig.DEBUG
-
-            fun Context.mustShowBatteryOptimizationDialog(): Boolean {
-                val sharedPrefs = getSharedPreferences(SHARED_PREFS, Activity.MODE_PRIVATE)
-                return sharedPrefs.getBoolean(FLAG_SHOW_BATTERY_OPTIMIZATION_DIALOG, true)
-            }
-
-            fun Context.updateShowBatteryOptimizationDialog(mustShowBatteryDialog: Boolean) {
-                val editableSharedPrefs = getSharedPreferences(SHARED_PREFS, Activity.MODE_PRIVATE).edit()
-                editableSharedPrefs.putBoolean(FLAG_SHOW_BATTERY_OPTIMIZATION_DIALOG, mustShowBatteryDialog).apply()
-            }
-        }
-
-        /**
-         * As long as it's false, the dialog will be displayed when user use a functionality needing background sync
-         *
-         * @see hasDoneNecessaryCheckbox
-         */
-        val hasDoneNecessary = MutableLiveData<Boolean>()
-
-        /**
-         * True if the IGNORE_BATTERY_OPTIMIZATIONS permission is given to the app
-         */
-        val isWhitelisted = MutableLiveData<Boolean>()
-
-        /**
-         * Requests battery permission when passed to true
-         *
-         * @see allowBackgroundSyncSwitch
-         */
-        val shouldBeWhitelisted = MutableLiveData<Boolean>()
-
-        /**
-         * Displays a warning card if true
-         *
-         * @see guideCard
-         */
-        val showManufacturerWarning = MutableLiveData<Boolean>()
-
-        fun showManufacturerWarning(hasBatteryPermission: Boolean) {
+    private fun showManufacturerWarning(hasBatteryPermission: Boolean) {
+        backgroundSyncPermissionsViewModel.apply {
             isWhitelisted.value = hasBatteryPermission
-            showManufacturerWarning.value = hasBatteryPermission && manufacturerWarning
+            guideCard.visibility = if (hasBatteryPermission && manufacturerWarning) View.VISIBLE else View.GONE
         }
+    }
+
+    companion object {
+        const val SUPPORT_FAQ_BACKGROUND_FUNCTIONING_URL = "https://faq.infomaniak.com/2685"
     }
 }
