+++ app/src/main/res/values-fr/strings.xml
@@ -82,12 +82,12 @@
     <string name="allYesterday">Hier</string>
     <string name="appSecurityDescription">Protégez l’accès à votre application en activant le verrouillage de votre téléphone.</string>
     <string name="appSecurityTitle">Verrouillage de l’application</string>
-    <string name="backgroundSyncCheckboxTitle">J\'ai fait le nécessaire</string>
+    <string name="backgroundSyncCheckboxTitle">J’ai fait le nécessaire</string>
     <string name="backgroundSyncDescription">Pour assurer le bon fonctionnement de kDrive, veuillez activer la synchronisation en arrière-plan sur votre téléphone</string>
     <string name="backgroundSyncOpenGuideButton">Ouvrir le guide</string>
-    <string name="backgroundSyncOpenGuideDescription">Il est possible que votre modèle de téléphone bloque la synchronisation en arrière-plan malgré l\'activation de l\'option. Veuillez suivre ce guide pour faire fonctionner la synchronisation. Sinon, vous devrez synchroniser manuellement vos fichiers depuis l\'application.</string>
-    <string name="backgroundSyncSwitchDescription">L\'application synchronisera automatiquement vos fichiers en arrière-plan.</string>
-    <string name="backgroundSyncSwitchTitle">Autoriser l\'application à fonctionner en arrière-plan</string>
+    <string name="backgroundSyncOpenGuideDescription">Il est possible que votre modèle de téléphone bloque la synchronisation en arrière-plan malgré l’activation de l’option. Veuillez suivre ce guide pour faire fonctionner la synchronisation. Sinon, vous devrez synchroniser manuellement vos fichiers depuis l’application.</string>
+    <string name="backgroundSyncSwitchDescription">L’application synchronisera automatiquement vos fichiers en arrière-plan.</string>
+    <string name="backgroundSyncSwitchTitle">Autoriser l’application à fonctionner en arrière-plan</string>
     <string name="backgroundSyncTitle">Synchronisation en arrière-plan</string>
     <string name="belongToAllCategoriesFilterDescription">Les résultats doivent appartenir à toutes les catégories sélectionnées</string>
     <string name="belongToOneCategoryFilterDescription">Les résultats doivent appartenir à une des catégories sélectionnées</string>
@@ -610,7 +610,7 @@
     <string name="sharedWithMeNoFile">Aucun fichier partagé avec moi</string>
     <string name="sharedWithMeTitle">Partagés avec moi</string>
     <string name="snackBarCategoryDeleted">Catégorie supprimée avec succès</string>
-    <string name="snackBarInvalidFileNameError">Le caractère \"%s\" du fichier %s n\'est pas pris en charge pour la mise hors ligne</string>
+    <string name="snackBarInvalidFileNameError">Le caractère \"%s\" du fichier %s n’est pas pris en charge pour la mise hors ligne</string>
     <plurals name="snackBarUploadError">
         <item quantity="one">Un fichier n’a pas pu être importé</item>
         <item quantity="other">%d fichiers n’ont pas pu être importés</item>
