+++ app/src/main/res/values-it/strings.xml
@@ -85,9 +85,9 @@
     <string name="backgroundSyncCheckboxTitle">Ho fatto ciò che era necessario</string>
     <string name="backgroundSyncDescription">Per garantire il corretto funzionamento di kDrive, abilita la sincronizzazione in background sul tuo telefono</string>
     <string name="backgroundSyncOpenGuideButton">Aprire la guida</string>
-    <string name="backgroundSyncOpenGuideDescription">È possibile che il tuo modello di telefono blocchi la sincronizzazione in background anche se l\'opzione è abilitata. Segui questa guida per far funzionare la sincronizzazione. Altrimenti, dovrete sincronizzare manualmente i vostri file dall\'applicazione.</string>
-    <string name="backgroundSyncSwitchDescription">L\'applicazione sincronizzerà automaticamente i vostri file in background.</string>
-    <string name="backgroundSyncSwitchTitle">Consentire l\'esecuzione dell\'applicazione in background</string>
+    <string name="backgroundSyncOpenGuideDescription">È possibile che il tuo modello di telefono blocchi la sincronizzazione in background anche se l’opzione è abilitata. Segui questa guida per far funzionare la sincronizzazione. Altrimenti, dovrete sincronizzare manualmente i vostri file dall’applicazione.</string>
+    <string name="backgroundSyncSwitchDescription">L’applicazione sincronizzerà automaticamente i vostri file in background.</string>
+    <string name="backgroundSyncSwitchTitle">Consentire l’esecuzione dell’applicazione in background</string>
     <string name="backgroundSyncTitle">Sincronizzazione in background</string>
     <string name="belongToAllCategoriesFilterDescription">I risultati devono appartenere a tutte le categorie selezionate</string>
     <string name="belongToOneCategoryFilterDescription">I risultati devono appartenere a una delle categorie selezionate</string>
@@ -610,7 +610,7 @@
     <string name="sharedWithMeNoFile">Nessun file condiviso con me</string>
     <string name="sharedWithMeTitle">Condivisi con me</string>
     <string name="snackBarCategoryDeleted">Categoria eliminata con successo</string>
-    <string name="snackBarInvalidFileNameError">Il carattere \"%s\" nel file %s non è supportato per l\'uso offline</string>
+    <string name="snackBarInvalidFileNameError">Il carattere \"%s\" nel file %s non è supportato per l’uso offline</string>
     <plurals name="snackBarUploadError">
         <item quantity="one">Non è stato possibile importare un file</item>
         <item quantity="other">%d file non possono essere importati</item>
